docs: move interfaces next to classes
