+++ docs/assets/js/search.js
@@ -1,3 +1,3 @@
 var typedoc = typedoc || {};
             typedoc.search = typedoc.search || {};
-            typedoc.search.data = {"kinds":{"1":"External module","4":"Enumeration","16":"Enumeration member","128":"Class","256":"Interface","512":"Constructor","1024":"Property"},"rows":[{"id":0,"kind":1,"name":"\"constants\"","url":"modules/_constants_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":1,"name":"\"enums\"","url":"modules/_enums_.html","classes":"tsd-kind-external-module"},{"id":2,"kind":4,"name":"NOTIFICATIONS_TYPE","url":"enums/_enums_.notifications_type.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"enums\""},{"id":3,"kind":16,"name":"PIPELINE_EXECUTION","url":"enums/_enums_.notifications_type.html#pipeline_execution","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":4,"kind":16,"name":"STAGE_EXECUTION","url":"enums/_enums_.notifications_type.html#stage_execution","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":5,"kind":16,"name":"ACTION_EXECUTION","url":"enums/_enums_.notifications_type.html#action_execution","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":6,"kind":16,"name":"NONE","url":"enums/_enums_.notifications_type.html#none","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":7,"kind":4,"name":"NOTIFICATIONS_TARGET","url":"enums/_enums_.notifications_target.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"enums\""},{"id":8,"kind":16,"name":"EMAIL","url":"enums/_enums_.notifications_target.html#email","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TARGET"},{"id":9,"kind":4,"name":"NOTIFICATIONS_DETAILS_TYPE","url":"enums/_enums_.notifications_details_type.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"enums\""},{"id":10,"kind":16,"name":"PIPELINE","url":"enums/_enums_.notifications_details_type.html#pipeline","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_DETAILS_TYPE"},{"id":11,"kind":16,"name":"STAGE","url":"enums/_enums_.notifications_details_type.html#stage","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_DETAILS_TYPE"},{"id":12,"kind":16,"name":"ACTION","url":"enums/_enums_.notifications_details_type.html#action","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_DETAILS_TYPE"},{"id":13,"kind":4,"name":"DOMAIN_NAME_REGISTRAR","url":"enums/_enums_.domain_name_registrar.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"enums\""},{"id":14,"kind":16,"name":"AWS","url":"enums/_enums_.domain_name_registrar.html#aws","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".DOMAIN_NAME_REGISTRAR"},{"id":15,"kind":1,"name":"\"interfaces\"","url":"modules/_interfaces_.html","classes":"tsd-kind-external-module"},{"id":16,"kind":256,"name":"IBackendDeploymentProps","url":"interfaces/_interfaces_.ibackenddeploymentprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":17,"kind":1024,"name":"notificationConfig","url":"interfaces/_interfaces_.ibackenddeploymentprops.html#notificationconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendDeploymentProps"},{"id":18,"kind":1024,"name":"pipelineConfig","url":"interfaces/_interfaces_.ibackenddeploymentprops.html#pipelineconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendDeploymentProps"},{"id":19,"kind":1024,"name":"description","url":"interfaces/_interfaces_.ibackenddeploymentprops.html#description","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendDeploymentProps"},{"id":20,"kind":1024,"name":"env","url":"interfaces/_interfaces_.ibackenddeploymentprops.html#env","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendDeploymentProps"},{"id":21,"kind":1024,"name":"stackName","url":"interfaces/_interfaces_.ibackenddeploymentprops.html#stackname","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendDeploymentProps"},{"id":22,"kind":1024,"name":"tags","url":"interfaces/_interfaces_.ibackenddeploymentprops.html#tags","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendDeploymentProps"},{"id":23,"kind":256,"name":"IFrontendDeploymentProps","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":24,"kind":1024,"name":"baseDomainName","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#basedomainname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":25,"kind":1024,"name":"hostedZoneName","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#hostedzonename","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":26,"kind":1024,"name":"certificateArn","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#certificatearn","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":27,"kind":1024,"name":"notificationConfig","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#notificationconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":28,"kind":1024,"name":"pipelineConfig","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#pipelineconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":29,"kind":1024,"name":"description","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#description","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":30,"kind":1024,"name":"env","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#env","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":31,"kind":1024,"name":"stackName","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#stackname","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":32,"kind":1024,"name":"tags","url":"interfaces/_interfaces_.ifrontenddeploymentprops.html#tags","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendDeploymentProps"},{"id":33,"kind":256,"name":"IBasePipelineProps","url":"interfaces/_interfaces_.ibasepipelineprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"interfaces\""},{"id":34,"kind":1024,"name":"artifactsSourceKey","url":"interfaces/_interfaces_.ibasepipelineprops.html#artifactssourcekey","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBasePipelineProps"},{"id":35,"kind":1024,"name":"environments","url":"interfaces/_interfaces_.ibasepipelineprops.html#environments","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBasePipelineProps"},{"id":36,"kind":1024,"name":"notificationsType","url":"interfaces/_interfaces_.ibasepipelineprops.html#notificationstype","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBasePipelineProps"},{"id":37,"kind":256,"name":"INotificationConfigProps","url":"interfaces/_interfaces_.inotificationconfigprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":38,"kind":1024,"name":"notificationsTargetConfig","url":"interfaces/_interfaces_.inotificationconfigprops.html#notificationstargetconfig","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".INotificationConfigProps"},{"id":39,"kind":256,"name":"INotificationsEmailTargetConfig","url":"interfaces/_interfaces_.inotificationsemailtargetconfig.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":40,"kind":1024,"name":"targetType","url":"interfaces/_interfaces_.inotificationsemailtargetconfig.html#targettype","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".INotificationsEmailTargetConfig"},{"id":41,"kind":1024,"name":"emailAddress","url":"interfaces/_interfaces_.inotificationsemailtargetconfig.html#emailaddress","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".INotificationsEmailTargetConfig"},{"id":42,"kind":256,"name":"IBackendEnvironment","url":"interfaces/_interfaces_.ibackendenvironment.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":43,"kind":1024,"name":"adminPermissions","url":"interfaces/_interfaces_.ibackendenvironment.html#adminpermissions","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":44,"kind":1024,"name":"privilegedMode","url":"interfaces/_interfaces_.ibackendenvironment.html#privilegedmode","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":45,"kind":1024,"name":"runtimeVariables","url":"interfaces/_interfaces_.ibackendenvironment.html#runtimevariables","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":46,"kind":1024,"name":"buildSpec","url":"interfaces/_interfaces_.ibackendenvironment.html#buildspec","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":47,"kind":1024,"name":"name","url":"interfaces/_interfaces_.ibackendenvironment.html#name","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendEnvironment"},{"id":48,"kind":1024,"name":"approvalRequired","url":"interfaces/_interfaces_.ibackendenvironment.html#approvalrequired","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendEnvironment"},{"id":49,"kind":256,"name":"IFrontendEnvironment","url":"interfaces/_interfaces_.ifrontendenvironment.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":50,"kind":1024,"name":"aliases","url":"interfaces/_interfaces_.ifrontendenvironment.html#aliases","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":51,"kind":1024,"name":"cloudfrontPriceClass","url":"interfaces/_interfaces_.ifrontendenvironment.html#cloudfrontpriceclass","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":52,"kind":1024,"name":"defaultRootObject","url":"interfaces/_interfaces_.ifrontendenvironment.html#defaultrootobject","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":53,"kind":1024,"name":"errorRootObject","url":"interfaces/_interfaces_.ifrontendenvironment.html#errorrootobject","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":54,"kind":1024,"name":"domainNameRegistrar","url":"interfaces/_interfaces_.ifrontendenvironment.html#domainnameregistrar","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":55,"kind":1024,"name":"name","url":"interfaces/_interfaces_.ifrontendenvironment.html#name","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendEnvironment"},{"id":56,"kind":1024,"name":"approvalRequired","url":"interfaces/_interfaces_.ifrontendenvironment.html#approvalrequired","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendEnvironment"},{"id":57,"kind":1,"name":"\"helpers/create-buildspec-with-credentials\"","url":"modules/_helpers_create_buildspec_with_credentials_.html","classes":"tsd-kind-external-module"},{"id":58,"kind":1,"name":"\"constructs/base-pipeline\"","url":"modules/_constructs_base_pipeline_.html","classes":"tsd-kind-external-module"},{"id":59,"kind":1,"name":"\"helpers/resolve-runtime-environments\"","url":"modules/_helpers_resolve_runtime_environments_.html","classes":"tsd-kind-external-module"},{"id":60,"kind":1,"name":"\"constructs/credential-store\"","url":"modules/_constructs_credential_store_.html","classes":"tsd-kind-external-module"},{"id":61,"kind":1,"name":"\"constructs/cdk-app-pipeline\"","url":"modules/_constructs_cdk_app_pipeline_.html","classes":"tsd-kind-external-module"},{"id":62,"kind":1,"name":"\"backend-deployment\"","url":"modules/_backend_deployment_.html","classes":"tsd-kind-external-module"},{"id":63,"kind":128,"name":"BackendDeployment","url":"classes/_backend_deployment_.backenddeployment.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"backend-deployment\""},{"id":64,"kind":512,"name":"constructor","url":"classes/_backend_deployment_.backenddeployment.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-overwrite","parent":"\"backend-deployment\".BackendDeployment"},{"id":65,"kind":1,"name":"\"constructs/web-distribution\"","url":"modules/_constructs_web_distribution_.html","classes":"tsd-kind-external-module"},{"id":66,"kind":1,"name":"\"constructs/frontend-cdn-pipeline\"","url":"modules/_constructs_frontend_cdn_pipeline_.html","classes":"tsd-kind-external-module"},{"id":67,"kind":1,"name":"\"frontend-deployment\"","url":"modules/_frontend_deployment_.html","classes":"tsd-kind-external-module"},{"id":68,"kind":128,"name":"FrontendDeployment","url":"classes/_frontend_deployment_.frontenddeployment.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"frontend-deployment\""},{"id":69,"kind":512,"name":"constructor","url":"classes/_frontend_deployment_.frontenddeployment.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-overwrite","parent":"\"frontend-deployment\".FrontendDeployment"}]};
+            typedoc.search.data = {"kinds":{"1":"External module","4":"Enumeration","16":"Enumeration member","128":"Class","256":"Interface","512":"Constructor","1024":"Property"},"rows":[{"id":0,"kind":1,"name":"\"enums\"","url":"modules/_enums_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":4,"name":"NOTIFICATIONS_TYPE","url":"enums/_enums_.notifications_type.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"enums\""},{"id":2,"kind":16,"name":"PIPELINE_EXECUTION","url":"enums/_enums_.notifications_type.html#pipeline_execution","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":3,"kind":16,"name":"STAGE_EXECUTION","url":"enums/_enums_.notifications_type.html#stage_execution","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":4,"kind":16,"name":"ACTION_EXECUTION","url":"enums/_enums_.notifications_type.html#action_execution","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":5,"kind":16,"name":"NONE","url":"enums/_enums_.notifications_type.html#none","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TYPE"},{"id":6,"kind":4,"name":"NOTIFICATIONS_TARGET","url":"enums/_enums_.notifications_target.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"enums\""},{"id":7,"kind":16,"name":"EMAIL","url":"enums/_enums_.notifications_target.html#email","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".NOTIFICATIONS_TARGET"},{"id":8,"kind":4,"name":"DOMAIN_NAME_REGISTRAR","url":"enums/_enums_.domain_name_registrar.html","classes":"tsd-kind-enum tsd-parent-kind-external-module","parent":"\"enums\""},{"id":9,"kind":16,"name":"AWS","url":"enums/_enums_.domain_name_registrar.html#aws","classes":"tsd-kind-enum-member tsd-parent-kind-enum","parent":"\"enums\".DOMAIN_NAME_REGISTRAR"},{"id":10,"kind":1,"name":"\"interfaces\"","url":"modules/_interfaces_.html","classes":"tsd-kind-external-module"},{"id":11,"kind":256,"name":"IBaseDeploymentProps","url":"interfaces/_interfaces_.ibasedeploymentprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"interfaces\""},{"id":12,"kind":1024,"name":"notificationConfig","url":"interfaces/_interfaces_.ibasedeploymentprops.html#notificationconfig","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBaseDeploymentProps"},{"id":13,"kind":1024,"name":"pipelineConfig","url":"interfaces/_interfaces_.ibasedeploymentprops.html#pipelineconfig","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBaseDeploymentProps"},{"id":14,"kind":256,"name":"IBasePipelineProps","url":"interfaces/_interfaces_.ibasepipelineprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"interfaces\""},{"id":15,"kind":1024,"name":"artifactsSourceKey","url":"interfaces/_interfaces_.ibasepipelineprops.html#artifactssourcekey","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBasePipelineProps"},{"id":16,"kind":1024,"name":"environments","url":"interfaces/_interfaces_.ibasepipelineprops.html#environments","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBasePipelineProps"},{"id":17,"kind":1024,"name":"notificationsType","url":"interfaces/_interfaces_.ibasepipelineprops.html#notificationstype","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBasePipelineProps"},{"id":18,"kind":256,"name":"INotificationConfigProps","url":"interfaces/_interfaces_.inotificationconfigprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":19,"kind":1024,"name":"notificationsTargetConfig","url":"interfaces/_interfaces_.inotificationconfigprops.html#notificationstargetconfig","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".INotificationConfigProps"},{"id":20,"kind":256,"name":"INotificationsEmailTargetConfig","url":"interfaces/_interfaces_.inotificationsemailtargetconfig.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":21,"kind":1024,"name":"targetType","url":"interfaces/_interfaces_.inotificationsemailtargetconfig.html#targettype","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".INotificationsEmailTargetConfig"},{"id":22,"kind":1024,"name":"emailAddress","url":"interfaces/_interfaces_.inotificationsemailtargetconfig.html#emailaddress","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".INotificationsEmailTargetConfig"},{"id":23,"kind":256,"name":"IBackendEnvironment","url":"interfaces/_interfaces_.ibackendenvironment.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":24,"kind":1024,"name":"adminPermissions","url":"interfaces/_interfaces_.ibackendenvironment.html#adminpermissions","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":25,"kind":1024,"name":"privilegedMode","url":"interfaces/_interfaces_.ibackendenvironment.html#privilegedmode","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":26,"kind":1024,"name":"runtimeVariables","url":"interfaces/_interfaces_.ibackendenvironment.html#runtimevariables","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":27,"kind":1024,"name":"buildSpec","url":"interfaces/_interfaces_.ibackendenvironment.html#buildspec","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IBackendEnvironment"},{"id":28,"kind":1024,"name":"name","url":"interfaces/_interfaces_.ibackendenvironment.html#name","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendEnvironment"},{"id":29,"kind":1024,"name":"approvalRequired","url":"interfaces/_interfaces_.ibackendenvironment.html#approvalrequired","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IBackendEnvironment"},{"id":30,"kind":256,"name":"IFrontendEnvironment","url":"interfaces/_interfaces_.ifrontendenvironment.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"interfaces\""},{"id":31,"kind":1024,"name":"aliases","url":"interfaces/_interfaces_.ifrontendenvironment.html#aliases","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":32,"kind":1024,"name":"cloudfrontPriceClass","url":"interfaces/_interfaces_.ifrontendenvironment.html#cloudfrontpriceclass","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":33,"kind":1024,"name":"defaultRootObject","url":"interfaces/_interfaces_.ifrontendenvironment.html#defaultrootobject","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":34,"kind":1024,"name":"errorRootObject","url":"interfaces/_interfaces_.ifrontendenvironment.html#errorrootobject","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":35,"kind":1024,"name":"domainNameRegistrar","url":"interfaces/_interfaces_.ifrontendenvironment.html#domainnameregistrar","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"interfaces\".IFrontendEnvironment"},{"id":36,"kind":1024,"name":"name","url":"interfaces/_interfaces_.ifrontendenvironment.html#name","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendEnvironment"},{"id":37,"kind":1024,"name":"approvalRequired","url":"interfaces/_interfaces_.ifrontendenvironment.html#approvalrequired","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"interfaces\".IFrontendEnvironment"},{"id":38,"kind":1,"name":"\"constants\"","url":"modules/_constants_.html","classes":"tsd-kind-external-module"},{"id":39,"kind":1,"name":"\"helpers/create-buildspec-with-credentials\"","url":"modules/_helpers_create_buildspec_with_credentials_.html","classes":"tsd-kind-external-module"},{"id":40,"kind":1,"name":"\"constructs/base-pipeline\"","url":"modules/_constructs_base_pipeline_.html","classes":"tsd-kind-external-module"},{"id":41,"kind":1,"name":"\"helpers/resolve-runtime-environments\"","url":"modules/_helpers_resolve_runtime_environments_.html","classes":"tsd-kind-external-module"},{"id":42,"kind":1,"name":"\"constructs/credential-store\"","url":"modules/_constructs_credential_store_.html","classes":"tsd-kind-external-module"},{"id":43,"kind":1,"name":"\"constructs/cdk-app-pipeline\"","url":"modules/_constructs_cdk_app_pipeline_.html","classes":"tsd-kind-external-module"},{"id":44,"kind":1,"name":"\"backend-deployment\"","url":"modules/_backend_deployment_.html","classes":"tsd-kind-external-module"},{"id":45,"kind":256,"name":"IBackendDeploymentProps","url":"interfaces/_backend_deployment_.ibackenddeploymentprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"backend-deployment\""},{"id":46,"kind":1024,"name":"notificationConfig","url":"interfaces/_backend_deployment_.ibackenddeploymentprops.html#notificationconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"backend-deployment\".IBackendDeploymentProps"},{"id":47,"kind":1024,"name":"pipelineConfig","url":"interfaces/_backend_deployment_.ibackenddeploymentprops.html#pipelineconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"backend-deployment\".IBackendDeploymentProps"},{"id":48,"kind":128,"name":"BackendDeployment","url":"classes/_backend_deployment_.backenddeployment.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"backend-deployment\""},{"id":49,"kind":512,"name":"constructor","url":"classes/_backend_deployment_.backenddeployment.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-overwrite","parent":"\"backend-deployment\".BackendDeployment"},{"id":50,"kind":1,"name":"\"constructs/web-distribution\"","url":"modules/_constructs_web_distribution_.html","classes":"tsd-kind-external-module"},{"id":51,"kind":1,"name":"\"constructs/frontend-cdn-pipeline\"","url":"modules/_constructs_frontend_cdn_pipeline_.html","classes":"tsd-kind-external-module"},{"id":52,"kind":1,"name":"\"frontend-deployment\"","url":"modules/_frontend_deployment_.html","classes":"tsd-kind-external-module"},{"id":53,"kind":256,"name":"IFrontendDeploymentProps","url":"interfaces/_frontend_deployment_.ifrontenddeploymentprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"frontend-deployment\""},{"id":54,"kind":1024,"name":"baseDomainName","url":"interfaces/_frontend_deployment_.ifrontenddeploymentprops.html#basedomainname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"frontend-deployment\".IFrontendDeploymentProps"},{"id":55,"kind":1024,"name":"hostedZoneName","url":"interfaces/_frontend_deployment_.ifrontenddeploymentprops.html#hostedzonename","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"frontend-deployment\".IFrontendDeploymentProps"},{"id":56,"kind":1024,"name":"certificateArn","url":"interfaces/_frontend_deployment_.ifrontenddeploymentprops.html#certificatearn","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"frontend-deployment\".IFrontendDeploymentProps"},{"id":57,"kind":1024,"name":"notificationConfig","url":"interfaces/_frontend_deployment_.ifrontenddeploymentprops.html#notificationconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"frontend-deployment\".IFrontendDeploymentProps"},{"id":58,"kind":1024,"name":"pipelineConfig","url":"interfaces/_frontend_deployment_.ifrontenddeploymentprops.html#pipelineconfig","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"frontend-deployment\".IFrontendDeploymentProps"},{"id":59,"kind":128,"name":"FrontendDeployment","url":"classes/_frontend_deployment_.frontenddeployment.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"frontend-deployment\""},{"id":60,"kind":512,"name":"constructor","url":"classes/_frontend_deployment_.frontenddeployment.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class tsd-is-overwrite","parent":"\"frontend-deployment\".FrontendDeployment"}]};
