+++ docs/classes/_backend_deployment_.backenddeployment.html
@@ -108,14 +108,14 @@
 					<a name="constructor" class="tsd-anchor"></a>
 					<h3>constructor</h3>
 					<ul class="tsd-signatures tsd-kind-constructor tsd-parent-kind-class tsd-is-overwrite">
-						<li class="tsd-signature tsd-kind-icon">new <wbr>Backend<wbr>Deployment<span class="tsd-signature-symbol">(</span>scope<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">App</span>, id<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span>, props<span class="tsd-signature-symbol">: </span><a href="../interfaces/_interfaces_.ibackenddeploymentprops.html" class="tsd-signature-type">IBackendDeploymentProps</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="_backend_deployment_.backenddeployment.html" class="tsd-signature-type">BackendDeployment</a></li>
+						<li class="tsd-signature tsd-kind-icon">new <wbr>Backend<wbr>Deployment<span class="tsd-signature-symbol">(</span>scope<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">App</span>, id<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span>, props<span class="tsd-signature-symbol">: </span><a href="../interfaces/_backend_deployment_.ibackenddeploymentprops.html" class="tsd-signature-type">IBackendDeploymentProps</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="_backend_deployment_.backenddeployment.html" class="tsd-signature-type">BackendDeployment</a></li>
 					</ul>
 					<ul class="tsd-descriptions">
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<p>Overrides Stack.__constructor</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/backend-deployment.ts#L12">package/resources/backend-deployment.ts:12</a></li>
+									<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/backend-deployment.ts#L27">package/resources/backend-deployment.ts:27</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -127,7 +127,7 @@
 									<h5>id: <span class="tsd-signature-type">string</span></h5>
 								</li>
 								<li>
-									<h5>props: <a href="../interfaces/_interfaces_.ibackenddeploymentprops.html" class="tsd-signature-type">IBackendDeploymentProps</a></h5>
+									<h5>props: <a href="../interfaces/_backend_deployment_.ibackenddeploymentprops.html" class="tsd-signature-type">IBackendDeploymentProps</a></h5>
 								</li>
 							</ul>
 							<h4 class="tsd-returns-title">Returns <a href="_backend_deployment_.backenddeployment.html" class="tsd-signature-type">BackendDeployment</a></h4>
@@ -170,6 +170,9 @@
 					</li>
 				</ul>
 				<ul class="after-current">
+					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
+						<a href="../interfaces/_backend_deployment_.ibackenddeploymentprops.html" class="tsd-kind-icon">IBackend<wbr>Deployment<wbr>Props</a>
+					</li>
 				</ul>
 			</nav>
 		</div>
