+++ docs/classes/_frontend_deployment_.frontenddeployment.html
@@ -108,14 +108,14 @@
 					<a name="constructor" class="tsd-anchor"></a>
 					<h3>constructor</h3>
 					<ul class="tsd-signatures tsd-kind-constructor tsd-parent-kind-class tsd-is-overwrite">
-						<li class="tsd-signature tsd-kind-icon">new <wbr>Frontend<wbr>Deployment<span class="tsd-signature-symbol">(</span>scope<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">App</span>, id<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span>, props<span class="tsd-signature-symbol">: </span><a href="../interfaces/_interfaces_.ifrontenddeploymentprops.html" class="tsd-signature-type">IFrontendDeploymentProps</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="_frontend_deployment_.frontenddeployment.html" class="tsd-signature-type">FrontendDeployment</a></li>
+						<li class="tsd-signature tsd-kind-icon">new <wbr>Frontend<wbr>Deployment<span class="tsd-signature-symbol">(</span>scope<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">App</span>, id<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span>, props<span class="tsd-signature-symbol">: </span><a href="../interfaces/_frontend_deployment_.ifrontenddeploymentprops.html" class="tsd-signature-type">IFrontendDeploymentProps</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="_frontend_deployment_.frontenddeployment.html" class="tsd-signature-type">FrontendDeployment</a></li>
 					</ul>
 					<ul class="tsd-descriptions">
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<p>Overrides Stack.__constructor</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/frontend-deployment.ts#L17">package/resources/frontend-deployment.ts:17</a></li>
+									<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/frontend-deployment.ts#L29">package/resources/frontend-deployment.ts:29</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -127,7 +127,7 @@
 									<h5>id: <span class="tsd-signature-type">string</span></h5>
 								</li>
 								<li>
-									<h5>props: <a href="../interfaces/_interfaces_.ifrontenddeploymentprops.html" class="tsd-signature-type">IFrontendDeploymentProps</a></h5>
+									<h5>props: <a href="../interfaces/_frontend_deployment_.ifrontenddeploymentprops.html" class="tsd-signature-type">IFrontendDeploymentProps</a></h5>
 								</li>
 							</ul>
 							<h4 class="tsd-returns-title">Returns <a href="_frontend_deployment_.frontenddeployment.html" class="tsd-signature-type">FrontendDeployment</a></h4>
@@ -170,6 +170,9 @@
 					</li>
 				</ul>
 				<ul class="after-current">
+					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
+						<a href="../interfaces/_frontend_deployment_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a>
+					</li>
 				</ul>
 			</nav>
 		</div>
