+++ docs/enums/_enums_.domain_name_registrar.html
@@ -86,7 +86,7 @@
 					<div class="tsd-signature tsd-kind-icon">AWS<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;aws&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in package/resources/enums.ts:19</li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/enums.ts#L22">package/resources/enums.ts:22</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -126,9 +126,6 @@
 					</li>
 				</ul>
 				<ul class="after-current">
-					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
-						<a href="_enums_.notifications_details_type.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>DETAILS_<wbr>TYPE</a>
-					</li>
 					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
 						<a href="_enums_.notifications_target.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>TARGET</a>
 					</li>
