+++ docs/enums/_enums_.notifications_target.html
@@ -86,7 +86,7 @@
 					<div class="tsd-signature tsd-kind-icon">EMAIL<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;email&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in package/resources/enums.ts:9</li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/enums.ts#L9">package/resources/enums.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -116,9 +116,6 @@
 				<ul class="before-current">
 					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
 						<a href="_enums_.domain_name_registrar.html" class="tsd-kind-icon">DOMAIN_<wbr>NAME_<wbr>REGISTRAR</a>
-					</li>
-					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
-						<a href="_enums_.notifications_details_type.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>DETAILS_<wbr>TYPE</a>
 					</li>
 				</ul>
 				<ul class="current">
