+++ docs/enums/_enums_.notifications_type.html
@@ -89,7 +89,7 @@
 					<div class="tsd-signature tsd-kind-icon">ACTION_<wbr>EXECUTION<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;actionExecution&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in package/resources/enums.ts:4</li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/enums.ts#L4">package/resources/enums.ts:4</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -99,7 +99,7 @@
 					<div class="tsd-signature tsd-kind-icon">NONE<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;none&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in package/resources/enums.ts:5</li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/enums.ts#L5">package/resources/enums.ts:5</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -109,7 +109,7 @@
 					<div class="tsd-signature tsd-kind-icon">PIPELINE_<wbr>EXECUTION<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;pipelineExecution&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in package/resources/enums.ts:2</li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/enums.ts#L2">package/resources/enums.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -119,7 +119,7 @@
 					<div class="tsd-signature tsd-kind-icon">STAGE_<wbr>EXECUTION<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;stageExecution&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in package/resources/enums.ts:3</li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/enums.ts#L3">package/resources/enums.ts:3</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -149,9 +149,6 @@
 				<ul class="before-current">
 					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
 						<a href="_enums_.domain_name_registrar.html" class="tsd-kind-icon">DOMAIN_<wbr>NAME_<wbr>REGISTRAR</a>
-					</li>
-					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
-						<a href="_enums_.notifications_details_type.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>DETAILS_<wbr>TYPE</a>
 					</li>
 					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
 						<a href="_enums_.notifications_target.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>TARGET</a>
