+++ docs/interfaces/_interfaces_.ibasepipelineprops.html
@@ -116,7 +116,7 @@
 					<div class="tsd-signature tsd-kind-icon">artifacts<wbr>Source<wbr>Key<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L52">package/resources/interfaces.ts:52</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L24">package/resources/interfaces.ts:24</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -126,7 +126,7 @@
 					<div class="tsd-signature tsd-kind-icon">environments<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">T</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L53">package/resources/interfaces.ts:53</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L25">package/resources/interfaces.ts:25</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -136,7 +136,7 @@
 					<div class="tsd-signature tsd-kind-icon">notifications<wbr>Type<span class="tsd-signature-symbol">:</span> <a href="../enums/_enums_.notifications_type.html" class="tsd-signature-type">NOTIFICATIONS_TYPE</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L57">package/resources/interfaces.ts:57</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L29">package/resources/interfaces.ts:29</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -174,10 +174,10 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_interfaces_.ibackenddeploymentprops.html" class="tsd-kind-icon">IBackend<wbr>Deployment<wbr>Props</a>
-					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
 						<a href="_interfaces_.ibackendenvironment.html" class="tsd-kind-icon">IBackend<wbr>Environment</a>
+					</li>
+					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
+						<a href="_interfaces_.ibasedeploymentprops.html" class="tsd-kind-icon">IBase<wbr>Deployment<wbr>Props</a>
 					</li>
 				</ul>
 				<ul class="current">
@@ -197,9 +197,6 @@
 					</li>
 				</ul>
 				<ul class="after-current">
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_interfaces_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a>
-					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
 						<a href="_interfaces_.ifrontendenvironment.html" class="tsd-kind-icon">IFrontend<wbr>Environment</a>
 					</li>
