+++ docs/interfaces/_interfaces_.ifrontendenvironment.html
@@ -123,7 +123,7 @@
 					<div class="tsd-signature tsd-kind-icon">aliases<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L107">package/resources/interfaces.ts:107</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L79">package/resources/interfaces.ts:79</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -134,7 +134,7 @@
 					<aside class="tsd-sources">
 						<p>Inherited from <a href="_interfaces_.ibackendenvironment.html">IBackendEnvironment</a>.<a href="_interfaces_.ibackendenvironment.html#approvalrequired">approvalRequired</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L135">package/resources/interfaces.ts:135</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L107">package/resources/interfaces.ts:107</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -153,7 +153,7 @@
 					<div class="tsd-signature tsd-kind-icon">cloudfront<wbr>Price<wbr>Class<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">PriceClass</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L111">package/resources/interfaces.ts:111</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L83">package/resources/interfaces.ts:83</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -170,7 +170,7 @@
 					<div class="tsd-signature tsd-kind-icon">default<wbr>Root<wbr>Object<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L115">package/resources/interfaces.ts:115</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L87">package/resources/interfaces.ts:87</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -187,7 +187,7 @@
 					<div class="tsd-signature tsd-kind-icon">domain<wbr>Name<wbr>Registrar<span class="tsd-signature-symbol">:</span> <a href="../enums/_enums_.domain_name_registrar.html" class="tsd-signature-type">DOMAIN_NAME_REGISTRAR</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L123">package/resources/interfaces.ts:123</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L95">package/resources/interfaces.ts:95</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -204,7 +204,7 @@
 					<div class="tsd-signature tsd-kind-icon">error<wbr>Root<wbr>Object<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L119">package/resources/interfaces.ts:119</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L91">package/resources/interfaces.ts:91</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -222,7 +222,7 @@
 					<aside class="tsd-sources">
 						<p>Inherited from <a href="_interfaces_.ibackendenvironment.html">IBackendEnvironment</a>.<a href="_interfaces_.ibackendenvironment.html#name">name</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L131">package/resources/interfaces.ts:131</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L103">package/resources/interfaces.ts:103</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -251,16 +251,13 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_interfaces_.ibackenddeploymentprops.html" class="tsd-kind-icon">IBackend<wbr>Deployment<wbr>Props</a>
-					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
 						<a href="_interfaces_.ibackendenvironment.html" class="tsd-kind-icon">IBackend<wbr>Environment</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="_interfaces_.ibasepipelineprops.html" class="tsd-kind-icon">IBase<wbr>Pipeline<wbr>Props</a>
+						<a href="_interfaces_.ibasedeploymentprops.html" class="tsd-kind-icon">IBase<wbr>Deployment<wbr>Props</a>
 					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_interfaces_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a>
+					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
+						<a href="_interfaces_.ibasepipelineprops.html" class="tsd-kind-icon">IBase<wbr>Pipeline<wbr>Props</a>
 					</li>
 				</ul>
 				<ul class="current">
