+++ docs/interfaces/_interfaces_.inotificationconfigprops.html
@@ -103,7 +103,7 @@
 					<div class="tsd-signature tsd-kind-icon">notifications<wbr>Target<wbr>Config<span class="tsd-signature-symbol">:</span> <a href="_interfaces_.inotificationsemailtargetconfig.html" class="tsd-signature-type">INotificationsEmailTargetConfig</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/eb1a1cd/package/resources/interfaces.ts#L64">package/resources/interfaces.ts:64</a></li>
+							<li>Defined in <a href="https://github.com/NextFaze/dootstrapper/blob/faf1b93/package/resources/interfaces.ts#L36">package/resources/interfaces.ts:36</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -132,16 +132,13 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_interfaces_.ibackenddeploymentprops.html" class="tsd-kind-icon">IBackend<wbr>Deployment<wbr>Props</a>
-					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
 						<a href="_interfaces_.ibackendenvironment.html" class="tsd-kind-icon">IBackend<wbr>Environment</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="_interfaces_.ibasepipelineprops.html" class="tsd-kind-icon">IBase<wbr>Pipeline<wbr>Props</a>
+						<a href="_interfaces_.ibasedeploymentprops.html" class="tsd-kind-icon">IBase<wbr>Deployment<wbr>Props</a>
 					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_interfaces_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a>
+					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
+						<a href="_interfaces_.ibasepipelineprops.html" class="tsd-kind-icon">IBase<wbr>Pipeline<wbr>Props</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
 						<a href="_interfaces_.ifrontendenvironment.html" class="tsd-kind-icon">IFrontend<wbr>Environment</a>
