+++ docs/modules/_enums_.html
@@ -70,7 +70,6 @@
 							<h3>Enumerations</h3>
 							<ul class="tsd-index-list">
 								<li class="tsd-kind-enum tsd-parent-kind-external-module"><a href="../enums/_enums_.domain_name_registrar.html" class="tsd-kind-icon">DOMAIN_<wbr>NAME_<wbr>REGISTRAR</a></li>
-								<li class="tsd-kind-enum tsd-parent-kind-external-module"><a href="../enums/_enums_.notifications_details_type.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>DETAILS_<wbr>TYPE</a></li>
 								<li class="tsd-kind-enum tsd-parent-kind-external-module"><a href="../enums/_enums_.notifications_target.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>TARGET</a></li>
 								<li class="tsd-kind-enum tsd-parent-kind-external-module"><a href="../enums/_enums_.notifications_type.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>TYPE</a></li>
 							</ul>
@@ -103,9 +102,6 @@
 				<ul class="before-current">
 					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
 						<a href="../enums/_enums_.domain_name_registrar.html" class="tsd-kind-icon">DOMAIN_<wbr>NAME_<wbr>REGISTRAR</a>
-					</li>
-					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
-						<a href="../enums/_enums_.notifications_details_type.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>DETAILS_<wbr>TYPE</a>
 					</li>
 					<li class=" tsd-kind-enum tsd-parent-kind-external-module">
 						<a href="../enums/_enums_.notifications_target.html" class="tsd-kind-icon">NOTIFICATIONS_<wbr>TARGET</a>
