+++ docs/modules/_frontend_deployment_.html
@@ -67,9 +67,15 @@
 				<section class="tsd-panel tsd-index-panel">
 					<div class="tsd-index-content">
 						<section class="tsd-index-section ">
-							<h3>frontend Classes</h3>
+							<h3>Classes</h3>
 							<ul class="tsd-index-list">
 								<li class="tsd-kind-class tsd-parent-kind-external-module"><a href="../classes/_frontend_deployment_.frontenddeployment.html" class="tsd-kind-icon">Frontend<wbr>Deployment</a></li>
+							</ul>
+						</section>
+						<section class="tsd-index-section ">
+							<h3>Interfaces</h3>
+							<ul class="tsd-index-list">
+								<li class="tsd-kind-interface tsd-parent-kind-external-module"><a href="../interfaces/_frontend_deployment_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a></li>
 							</ul>
 						</section>
 					</div>
@@ -100,6 +106,9 @@
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-parent-kind-external-module">
 						<a href="../classes/_frontend_deployment_.frontenddeployment.html" class="tsd-kind-icon">Frontend<wbr>Deployment</a>
+					</li>
+					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
+						<a href="../interfaces/_frontend_deployment_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a>
 					</li>
 				</ul>
 			</nav>
