+++ docs/modules/_interfaces_.html
@@ -69,10 +69,9 @@
 						<section class="tsd-index-section ">
 							<h3>Interfaces</h3>
 							<ul class="tsd-index-list">
-								<li class="tsd-kind-interface tsd-parent-kind-external-module"><a href="../interfaces/_interfaces_.ibackenddeploymentprops.html" class="tsd-kind-icon">IBackend<wbr>Deployment<wbr>Props</a></li>
 								<li class="tsd-kind-interface tsd-parent-kind-external-module"><a href="../interfaces/_interfaces_.ibackendenvironment.html" class="tsd-kind-icon">IBackend<wbr>Environment</a></li>
+								<li class="tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter"><a href="../interfaces/_interfaces_.ibasedeploymentprops.html" class="tsd-kind-icon">IBase<wbr>Deployment<wbr>Props</a></li>
 								<li class="tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter"><a href="../interfaces/_interfaces_.ibasepipelineprops.html" class="tsd-kind-icon">IBase<wbr>Pipeline<wbr>Props</a></li>
-								<li class="tsd-kind-interface tsd-parent-kind-external-module"><a href="../interfaces/_interfaces_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a></li>
 								<li class="tsd-kind-interface tsd-parent-kind-external-module"><a href="../interfaces/_interfaces_.ifrontendenvironment.html" class="tsd-kind-icon">IFrontend<wbr>Environment</a></li>
 								<li class="tsd-kind-interface tsd-parent-kind-external-module"><a href="../interfaces/_interfaces_.inotificationconfigprops.html" class="tsd-kind-icon">INotification<wbr>Config<wbr>Props</a></li>
 								<li class="tsd-kind-interface tsd-parent-kind-external-module"><a href="../interfaces/_interfaces_.inotificationsemailtargetconfig.html" class="tsd-kind-icon">INotifications<wbr>Email<wbr>Target<wbr>Config</a></li>
@@ -105,16 +104,13 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="../interfaces/_interfaces_.ibackenddeploymentprops.html" class="tsd-kind-icon">IBackend<wbr>Deployment<wbr>Props</a>
-					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
 						<a href="../interfaces/_interfaces_.ibackendenvironment.html" class="tsd-kind-icon">IBackend<wbr>Environment</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="../interfaces/_interfaces_.ibasepipelineprops.html" class="tsd-kind-icon">IBase<wbr>Pipeline<wbr>Props</a>
+						<a href="../interfaces/_interfaces_.ibasedeploymentprops.html" class="tsd-kind-icon">IBase<wbr>Deployment<wbr>Props</a>
 					</li>
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="../interfaces/_interfaces_.ifrontenddeploymentprops.html" class="tsd-kind-icon">IFrontend<wbr>Deployment<wbr>Props</a>
+					<li class=" tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
+						<a href="../interfaces/_interfaces_.ibasepipelineprops.html" class="tsd-kind-icon">IBase<wbr>Pipeline<wbr>Props</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
 						<a href="../interfaces/_interfaces_.ifrontendenvironment.html" class="tsd-kind-icon">IFrontend<wbr>Environment</a>
